pyproject.toml
[tool.poetry]
name = "course"
version = "0.1.0"
description = ""
authors = ["jenni"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
SQLAlchemy = "^1.3.22"
black = "^20.8b1"
flake8 = "^3.8.4"
psycopg2 = "^2.8.6"
PyJWT = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
